ipn/store: moved it into ipnstate package
Initially, I believed it would be interesting to have a dedicated
package to develop the StateStore interface and its
implementation.

After some discussions with @bradfitz, we figured that it
would make more sense to actually move it as part of the
ipn/ipnstate one.

The main motivation behind this change was to get
started to work onto new StateStore implementation.
I had in mind a couple of alternatives like leveraging:

- AWS Systems Manager Parameter Store
- Kubernetes custom resources

Which could potentially allow end-users running on
WS or K8S to avoid mounting a 1GB volume to store a
tiny JSON file.

This change is not a prereq nor a blocker but I felt
it would be cleaner to implement it in such fashion.

Signed-off-by: Maxime VISONNEAU <maxime.visonneau@gmail.com>