derp: log highest bandwidth ip
Adds metrics for reporting which ips are using the most bandwidth, which would hopefully give
insight into the locations using the most derp bandwidth. I wasn't sure what N to make it, so I
left it as 100, but probably makes sense to scale up further depending on how many users there
are, possibly dynamically in the future.

Signed-off-by: julianknodt <julianknodt@gmail.com>