net/tstun: remove multi-case selects from hot code
Every TUN Read went through several multi-case selects.
We know from past experience with wireguard-go that these are slow
and cause scheduler churn.

The selects served two purposes: they separated errors from data and
handled shutdown. The first is fairly easy to replace by sending
errors and data over a single channel. The second, less so.

The new strategy for managing shutdown is a lot less graceful.

At the beginning of each Read call and poll loop, we check whether the
device is closed and exit if so.
We thus only need worry about a single execution.

Read and poll interact with two channels: bufferC and outbound.
Close sends sentinel values on those channels,
ensuring that Read and poll won't block indefinitely on receives.

We also need to ensure that they won't block indefinitely on channel sends.
We arrange for there to be a limit on the number of concurrent
senders on bufferC and outbound, and give each of them both a buffer
large enough that no send to it can ever block.

One behavior change of note: If Close and InjectOutbound are
called concurrently, it is now possible for InjectOutbound
to return a nil error, yet have that packet never be Read.
In practice, I don't think that this is a problem.

Updates tailscale/corp#212

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>