net/tstun: eliminate per-packet time.Now calls
We were calling time.Now on every single packet.
We don't need that level of precision;
it is only used to track how long the TUN device has been idle,
to a one second granularity.

Replace this with a loop that checks for activity every second
and increments a counter if so.

It is hard to measure precisely, but this increases linux VM to
darwin host throughput by as much as 10%.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>