net/portmapper: Add discovery to upnp
Previously, it was just a brittle check for a specific port on my own network.
It seems that routers are free to pick whatever port they'd like, so if we'd really like to
support any router it's necessary to have discovery. This is the smallest possible change
to add it in.

Signed-off-by: julianknodt <julianknodt@gmail.com>