tstest/integration/vms: mv into own fns
While reading through these tests, I found it really difficult to grok the dependencies of each
one, and if any of them modify the environment in some way. This makes it clear that each test
is running in its own scope, and what is exactly necessary for each test, as well as making the
entry point cleaner and easier to add into for more tests.

Signed-off-by: julianknodt <julianknodt@gmail.com>